requirements.txt
mkdocs
mkdocstrings
mkdocs-material
pyyaml
mkdocs-material-extensions